Added clean for ios

+++ build.sh
@@ -29,7 +29,11 @@ elif [[ "${1:-}" == "ios" ]]; then
 	bash _ops/build.ios.sh
 
 elif [[ "${1:-}" == "clean" ]]; then
-	docker run -e TMDB_KEY -v ~/.pub-cache:/pub-cache -e PUB_CACHE=/pub-cache -v `pwd`:/src -w /src --rm fmv_tools bash _ops/clean.sh
+	if type docker; then
+		docker run -e TMDB_KEY -v ~/.pub-cache:/pub-cache -e PUB_CACHE=/pub-cache -v `pwd`:/src -w /src --rm fmv_tools bash _ops/clean.sh
+	else
+		PUB_CACHE=~/.pub-cache bash _ops/clean.sh
+	fi
 
 else
 	echo "---"
